apply library file layout
