+++ types/node-os-utils/lib/index.d.ts
@@ -1,0 +1,33 @@
+import Cpu from './cpu';
+import Drive from './drive';
+import Mem from './mem';
+import NetStat from './netstat';
+import OpenFiles from './openfiles';
+import Os from './os';
+import OsCmd from './oscmd';
+import Proc from './proc';
+import Users from './users';
+
+export const cpu: Cpu;
+export const drive: Drive;
+export const mem: Mem;
+export const netstat: NetStat;
+export const openfiles: OpenFiles;
+export const os: Os;
+export const oscmd: OsCmd;
+export const proc: Proc;
+export const users: Users;
+
+export let options: { NOT_SUPPORTED_VALUE: string, INTERVAL: number };
+
+export function exec(command: string): () => Promise<string>;
+
+export * from './cpu'
+export * from './drive'
+export * from './mem'
+export * from './netstat'
+export * from './openfiles'
+export * from './os'
+export * from './oscmd'
+export * from './proc'
+export * from './users'
