+++ types/node-os-utils/lib/openfiles.d.ts
@@ -1,0 +1,3 @@
+export default class OpenFiles {
+    openFd(): Promise<number>;
+}
