+++ types/node-os-utils/lib/proc.d.ts
@@ -1,0 +1,5 @@
+export default class Proc {
+    totalProcesses(): Promise<number | string>;
+
+    zombieProcesses(): Promise<number | string>;
+}
