+++ types/node-os-utils/lib/users.d.ts
@@ -1,0 +1,3 @@
+export default class Users {
+    openedCount(): Promise<number | string>;
+}
